pyproject.toml
[tool.poetry]
name = "preview-link"
version = "0.1.0"
description = ""
authors = ["hoanghuumanh54 <hoanghuumanh54@gmail.com>"]
readme = "README.md"
packages = [{include = "preview_link"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
Pillow = "^9.4.0"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
